Cleans up encoders.
