+++ integration/simulation/validate_chain.go
@@ -5,6 +5,8 @@ import (
 	"sync"
 	"testing"
 
+	"github.com/obscuronet/obscuro-playground/go/common/log"
+
 	"github.com/obscuronet/obscuro-playground/go/enclave/bridge"
 
 	"github.com/obscuronet/obscuro-playground/go/rpcclientlib"
@@ -157,7 +159,10 @@ func ExtractDataFromEthereumChain(startBlock *types.Block, endBlock *types.Block
 				deposits = append(deposits, tx.Hash())
 				totalDeposited += l1tx.Amount
 			case *ethadapter.L1RollupTx:
-				r := common.DecodeRollupOrPanic(l1tx.Rollup)
+				r, err := common.DecodeRollup(l1tx.Rollup)
+				if err != nil {
+					log.Panic("could not decode rollup. Cause: %s", err)
+				}
 				rollups = append(rollups, r.Hash())
 				if node.IsBlockAncestor(block, r.Header.L1Proof) {
 					// only count the rollup if it is published in the right branch
